extract sysinfo from python http stager
